requirements.txt
aioredis
fastapi
flask
google
google-auth-oauthlib
httpx
lanarky
langchain
loguru
openai
passlib
pathlib
pinecone-client
pydantic
pymongo
python-dotenv
python-jose
python-multipart
python-multipart
redis
requests
sqlalchemy
starlette
starlette-session
streamlit
tiktoken
uvicorn
streamlit_extras
google-auth
httpx_oauth